Move content store implementation
Move the filesystem content store implementation to a
subpackage of content.
