+++ content/content.go
@@ -3,18 +3,9 @@ package content
 import (
 	"context"
 	"io"
-	"sync"
 	"time"
 
 	"github.com/opencontainers/go-digest"
-)
-
-var (
-	bufPool = sync.Pool{
-		New: func() interface{} {
-			return make([]byte, 1<<20)
-		},
-	}
 )
 
 type Provider interface {
