+++ content/fs/store_linux.go
@@ -1,4 +1,4 @@
-package content
+package fs
 
 import (
 	"os"
